+++ semcore/radio/src/Radio.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import createComponent, { Component, CONTEXT_COMPONENT, sstyled, Root } from '@semcore/core';
-import { Flex, Box, InvalidStatePattern } from '@semcore/flex-box';
+import { Flex, Box, InvalidStateBox } from '@semcore/flex-box';
 import assignProps, { callAllEventHandlers } from '@semcore/utils/lib/assignProps';
 import keyboardFocusEnhance from '@semcore/utils/lib/enhances/keyboardFocusEnhance';
 import resolveColorEnhance from '@semcore/utils/lib/enhances/resolveColorEnhance';
@@ -251,7 +251,7 @@ Control.displayName = 'Control';
 
 const RadioMark = (props) => {
   const SValue = Root;
-  const SInvalidPattern = InvalidStatePattern;
+  const SInvalidPattern = InvalidStateBox;
   const { theme, styles, resolveColor, state, checked } = props;
 
   return sstyled(styles)(